refactor: extract mixins and attrs utils to attrs_utils.py
