+++ interactions/api/models/gw.py
@@ -2,20 +2,18 @@ from datetime import datetime
 from typing import Any, List, Optional, Union
 
 from ...client.models.component import ActionRow, Button, SelectMenu, _build_components
-from .channel import Channel, ThreadMember
-from .member import Member
-from .message import Embed, Emoji, Message, MessageInteraction, Sticker
-from .misc import (
+from .attrs_utils import (
     MISSING,
     ClientSerializerMixin,
-    ClientStatus,
     DictSerializerMixin,
-    File,
-    Snowflake,
     convert_list,
     define,
     field,
 )
+from .channel import Channel, ThreadMember
+from .member import Member
+from .message import Embed, Emoji, Message, MessageInteraction, Sticker
+from .misc import ClientStatus, File, Snowflake
 from .presence import PresenceActivity
 from .role import Role
 from .team import Application
