+++ interactions/client/models/misc.py
@@ -1,9 +1,10 @@
 from typing import Any, Dict, List, Optional
 
+from ...api.models.attrs_utils import DictSerializerMixin, convert_list, define, field
 from ...api.models.channel import Channel
 from ...api.models.member import Member
 from ...api.models.message import Attachment, Message
-from ...api.models.misc import DictSerializerMixin, Snowflake, convert_list, define, field
+from ...api.models.misc import Snowflake
 from ...api.models.role import Role
 from ...api.models.user import User
 from ..enums import ApplicationCommandType, ComponentType, InteractionType, PermissionType
